RavenDB-18701 - ThrowDiskFullException method moved to ExceptionHelper class.
